+++ lib/dart_helper_utils.dart
@@ -1,5 +1,1 @@
-/// The Dart Helper Package offers various extensions and helper
-/// methods that can make development more efficient.
-library dart_helper_utils;
-
 export 'src/src.dart';