+++ lib/src/extensions/duration.dart
@@ -1,8 +1,14 @@
 import 'dart:async';
 
-import 'date.dart';
+import 'package:dart_helper_utils/dart_helper_utils.dart';
 
-extension DurationExt on Duration {
+extension DHUDurationExt on Duration {
+  /// Adds the Duration to the current DateTime and returns a DateTime in the future
+  DateTime get fromNow => DateTime.now() + this;
+
+  /// Subtracts the Duration from the current DateTime and returns a DateTime in the past
+  DateTime get ago => DateTime.now() - this;
+
   /// Utility to delay some callback (or code execution).
   ///
   /// Sample:
@@ -12,12 +18,6 @@ extension DurationExt on Duration {
   ///   }
   ///
   ///```
-  Future<T> delay<T>([FutureOr<T> Function()? callback]) async =>
-      Future<T>.delayed(this, callback);
-
-  /// Adds the Duration to the current DateTime and returns a DateTime in the future
-  DateTime get fromNow => DateTime.now() + this;
-
-  /// Subtracts the Duration from the current DateTime and returns a DateTime in the past
-  DateTime get ago => DateTime.now() - this;
+  Future<T> delayed<T>([FutureOr<T> Function()? computation]) async =>
+      Future<T>.delayed(this, computation);
 }