+++ lib/src/extensions/extensions.dart
@@ -1,7 +1,10 @@
+export 'bool.dart';
 export 'date.dart';
 export 'duration.dart';
 export 'iterable.dart';
 export 'map.dart';
 export 'numbers.dart';
 export 'objects.dart';
-export 'string_ext.dart';
+export 'set.dart';
+export 'string_extensions/string_extensions.dart';
+export 'uri_extension.dart';