+++ lib/src/extensions/set.dart
@@ -1,0 +1,5 @@
+extension DHUNullableSetExt<E> on Set<E>? {
+  bool get isEmptyOrNull => this == null || this!.isEmpty;
+
+  bool get isNotEmptyOrNull => !isEmptyOrNull;
+}