+++ lib/src/src.dart
@@ -1,2 +1,5 @@
+export 'package:dart_helper_utils/dart_helper_utils.dart';
+
+export 'enums/enums.dart';
 export 'extensions/extensions.dart';
-export 'helper_methods/helper_methods.dart';
+export 'other_utils/other_utils.dart';